All Order routes working. Took out the activeOrder approach. All routes rely on session.order for both logged in users and guests.
If a user logs in, has an empty session.order but has an old & active order in the db that will be fetched and placed on session.order.

If a logged in user modifies their order session.order is updated and then that data is sent to persist in the db.

When a user logs out session.order is deleted.

Routes assume post data will = {"product": {"id": 4}, "total": "20.00"}